+++ src/lib/forms-schema/index.ts
@@ -68,6 +68,7 @@ const SubmissionEventsSchema = Joi.object().keys({
       'WESTPAC_QUICK_WEB',
       'CIVICA_CRM',
       'SCHEDULING',
+      'FRESHDESK_CREATE_TICKET',
     ),
   configuration: Joi.object()
     .required()
@@ -205,6 +206,27 @@ const SubmissionEventsSchema = Joi.object().keys({
         nameElementId: Joi.string().guid(),
         emailElementId: Joi.string().guid(),
         emailDescription: Joi.string(),
+      }),
+    })
+    .when('type', {
+      is: 'FRESHDESK_CREATE_TICKET',
+      then: Joi.object().keys({
+        mapping: Joi.array().items(
+          Joi.object().keys({
+            freshdeskFieldName: Joi.string().required(),
+            type: Joi.string().valid('FORM_ELEMENT', 'VALUE').required(),
+            formElementId: Joi.when('type', {
+              is: 'FORM_ELEMENT',
+              then: Joi.string().uuid().required(),
+              otherwise: Joi.any().strip(),
+            }),
+            value: Joi.when('type', {
+              is: 'VALUE',
+              then: Joi.alternatives().try(Joi.string(), Joi.number()),
+              otherwise: Joi.any().strip(),
+            }),
+          }),
+        ),
       }),
     }),
   conditionallyExecute: Joi.bool().default(false),