+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added submission event for Freshdesk ticket creation
+
 ## [0.18.2] - 2021-11-24
 
 ### Changed